+++ projet/src/cloth/cloth.hpp
@@ -19,8 +19,8 @@ struct cloth_structure
     cgp::numarray<cgp::uint3> triangle_connectivity;
 
     // The size of the cloth
-    int lenght_x;
-    int lenght_y;
+    float lenght_x;
+    float lenght_y;
 
     // The texture of the cloth
     cgp::opengl_texture_image_structure texture;
@@ -31,7 +31,7 @@ struct cloth_structure
     float mu = 15.0f;        // damping parameter
     
     
-    void initialize(int N_samples_edge, std::vector<vec3> pos, int x_lenght, int y_lenght);  // Initialize a square flat cloth
+    void initialize(int N_samples_edge, std::vector<vec3> pos, float x_lenght, float y_lenght);  // Initialize a square flat cloth
     void update_normal();       // Call this function every time the cloth is updated before its draw
     int N_samples_x() const;      // Number of vertex along x dimension of the grid
     int N_samples_y() const;      // Number of vertex along y dimension of the grid
